% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utilities.R
\name{case_study_nice_names}
\alias{case_study_nice_names}
\title{Translate var names to readable names}
\usage{
case_study_nice_names(varname)
}
\description{
Translate var names to readable names
}
\keyword{internal}
