% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-results-utils.R
\name{classify_effect}
\alias{classify_effect}
\title{Classify estimates based on confidence limits}
\usage{
classify_effect(lbcat, ubcat)
}
\description{
Classify estimates based on confidence limits
}
\keyword{internal}
