% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-results-utils.R
\name{conf_limit_category}
\alias{conf_limit_category}
\title{Classify confidence limits with respect to threshold value and zero}
\usage{
conf_limit_category(x, threshold_null_or)
}
\description{
Classify confidence limits with respect to threshold value and zero
}
\keyword{internal}
