% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-data-augment.R
\name{cox_data_augment}
\alias{cox_data_augment}
\title{Cox Regression with option of semi-Bayesian estimation through augmented data}
\usage{
cox_data_augment(t0_var, t1_var, y_var, x_var, prior_hr_ub_95 = Inf,
  data = NULL, normal = 0, return_as = FALSE)
}
\arguments{
\item{t0_var}{(string) variable name for observation start time}

\item{t1_var}{(string) variable name for observation end time}

\item{y_var}{(string) variable name for event indicator (1/0)}

\item{x_var}{(string) variable name for predictor}

\item{prior_hr_ub_95}{(numeric) upper 95% confidence limit for prior hazard
ratio. If infinite, will estimate with MLE method. If NA, will use Firth
method.}

\item{data}{dataset for estimation}

\item{normal}{(Boolean) use a normal distribution? If FALSE, then use a log-F distribution
for the log-hazard ratio}

\item{return_as}{(Boolean) return parameter values `A`
(number of cases represented by each prior data record) and
`S` (scale factor)?
If TRUE, then return them along with the primary results and the dataset.}
}
\description{
Estimates either standard Cox regression (MLE), Firth method, or
approximate Bayesian estimation given an upper 95% confidence limit for
the prior hazard ratio.
}
\references{
Sullivan SG, Greenland S. Bayesian regression in SAS software. Int J Epidemiol. 2013;42:308-317.
}
\keyword{internal}
