% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-model-wrapper.R
\name{cox_model_wrapper}
\alias{cox_model_wrapper}
\title{Wrapper for Cox Model estimation}
\usage{
cox_model_wrapper(data, n, u, r, var, error_dir, scale_x = FALSE)
}
\arguments{
\item{data}{(tbl) simluated or case study dataset from `generate_data_wrapper`}

\item{n}{(integer) number of events to include (will use the first `n` events)}

\item{u}{prior 95-percent upper bound on hazard ratio.}

\item{r}{(Boolean) use a normally distributed (TRUE) or F-distributed (FALSE)
prior on the hazard ratios?}

\item{var}{(string) predictor variable name}

\item{error_dir}{output directory - really just to save errors}

\item{scale_x}{(Boolean) scale X to have a standard deviation of 1 (TRUE)
or leave the current standard deviation (FALSE). (Note: X is always centered,
the option here is just whether to modify the standard deviation)}
}
\value{
tibble with confidence intervals and details for specified model
}
\description{
Wrapper for Cox Model estimation
}
