% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-named-beta.R
\name{create_named_beta}
\alias{create_named_beta}
\title{Create named vector of true coefficients}
\usage{
create_named_beta(sim = TRUE, param_row = NA)
}
\arguments{
\item{sim}{(boolean) TRUE = simulated data, FALSE = case study data}

\item{param_row}{(integer) which row of the parameter dataset to use
(only relevant for `sim = TRUE`)}
}
\value{
named numeric vector of coefficients
}
\description{
Create named vector of true coefficients
}
