% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-params.R
\name{create_params}
\alias{create_params}
\title{Create and save parameter table.}
\usage{
create_params(save_dir, n_subj, n_vars, n_imp = 0, freq = NA,
  var_types = NA, beta_imp = NA, wb_shape = 2, max_n_events = 80,
  sim = TRUE, nosave = FALSE, round_event_times = FALSE)
}
\arguments{
\item{save_dir}{location to save the parameter tbl}

\item{n_subj}{number of subjects}

\item{n_vars}{number of predictor variables}

\item{n_imp}{Number of "important" predictors (with nonzero coefficients)}

\item{freq}{for binary predictor variables, frequencies}

\item{var_types}{(string) either one character giving the type for all
predictor variables, or a string of length n_vars.
"c" = continuous, "b" = binary. If NA, then parameters will be generated for
both binary and continuous settings.}

\item{beta_imp}{Coefficients (hazard ratios) for important predictors}

\item{wb_shape}{Shape of Weibull distribution for event times}

\item{max_n_events}{(integer) Maximum number of events to simulate for any individual.}

\item{sim}{(Boolean) TRUE = simulated data, FALSE = case study data}

\item{nosave}{(Boolean) Don't save output, just return it (TRUE) or save it
as rds in the save_dir (FALSE, default)}

\item{round_event_times}{(Boolean) TRUE = round event times up to the nearest
integer, FALSE = keep exact event times}
}
\value{
parameter table (if `nosave` = TRUE) or simply the number of rows of
the table (if `nosave` = FALSE). The parameter table has the following columns:
\describe{
  \item{n.subj}{(integer) Number of participants}
  \item{n.events.max}{(integer) Number of outcome events simulated per participant.}
  \item{median.days.btwn.events}{(numeric) Target median inter-event time (days)}
  \item{wb.shape}{(numeric) Weibull distribution shape parameter for baseline hazard}
  \item{round_event_times}{(Boolean) whether to round up to next integer day (TRUE) or keep exact times (FALSE)}
  \item{freq}{(numeric) target frequency of dichotomous exposure}
  \item{n.vars}{(integer) number of predictor variables}
  \item{n.imp}{(integer) number of 'important' predictor variables with non-zero log-HR}
  \item{var.types}{(string of length `n.vars`, giving types of each predictor variable ('b' = binary/dichotomous, 'c' = continuous)}
  \item{beta}{(numeric vector of length `n.vars`) true log-HR values}
  \item{rho}{(numeric) autocorrelation coefficient (unused)}
  \item{p}{(integer) parameter table row ID}
 }
}
\description{
Create and save parameter table.
}
