% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-params.R
\name{estimation_params}
\alias{estimation_params}
\title{Create table of model estimation parameters.}
\usage{
estimation_params(sim, mv)
}
\arguments{
\item{sim}{(Boolean) TRUE if using simulated data, FALSE if real data}

\item{mv}{(Boolean) TRUE if multivariable, FALSE if univariable}
}
\value{
tbl with columns specifying estimation parameters:
\describe{
  \item{n.events}{number of events (cumulative) to include in estimation}
  \item{prior_hr_ub}{prior upper bound on hazard ratio. For proper normal or log-F priors on the log-hazard ratio, this is numeric. `Inf` indicates maximum partial likelihood estimation and `NA` indicates Firth estimation.}
  \item{prior_normal}{Boolean indicating whether the prior on the log-hazard ratio is normal (1) or log-F (0)}
}
}
\description{
Model estimation parameters include the number of events,
the 95% prior upper bound on the hazard ratio, and whether the
prior distribution is normal or log-F.
}
