% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-cis.R
\name{f}
\alias{f}
\title{Difference between profile log-likelihood and CI cutoff value.}
\usage{
f(initial_estimate, predictor_var_names, which_var, model_obj, maxloglik,
  CI, data)
}
\arguments{
\item{initial_estimate}{initial estimate}

\item{predictor_var_names}{(character vector) names of predictor variables}

\item{which_var}{(integer) index (into `predictor_var_names` of the variable
for which to create the confidence interval}

\item{model_obj}{original model object}

\item{maxloglik}{maximum log-likelihood from original model}

\item{CI}{(numeric) nominal confidence interval coverage level, strictly
between 0 and 1}
}
\description{
The zeros of this function are the confidence limits for
the estimated parameter.
}
\references{
Much of this code was taken from
Patterson, David. "Profile Likelihood Confidence Intervals for GLM's."
David Patterson, Math, U. of Montana.
http://www.math.umt.edu/patterson/ProfileLikelihoodCI.pdf Accessed 2020-01-09.
}
\keyword{internal}
