% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-cs.R
\name{figures_cs}
\alias{figures_cs}
\title{Plot figures for case study}
\usage{
figures_cs(run_dir)
}
\arguments{
\item{run_dir}{output directory for simulation & case study results}
}
\description{
Plot figures for case study
}
