% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mv.R
\name{figures_mv}
\alias{figures_mv}
\title{Plot figures for multivariable simulation results}
\usage{
figures_mv(run_dir)
}
\arguments{
\item{run_dir}{output directory for simulation & case study results}
}
\description{
Plot figures for multivariable simulation results
}
