% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-supp.R
\name{figures_supp}
\alias{figures_supp}
\title{Plot figures for supplement}
\usage{
figures_supp(run_dir)
}
\arguments{
\item{run_dir}{output directory for simulation & case study results}
}
\description{
Plot figures for supplement
}
