% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-univ.R
\name{figures_univ}
\alias{figures_univ}
\title{Plot figures for univariable results}
\usage{
figures_univ(run_dir)
}
\arguments{
\item{run_dir}{output directory for simulation & case study results}
}
\description{
Plot figures for univariable results
}
