% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-cis.R
\name{find_zero}
\alias{find_zero}
\title{Find zeros of a concave function}
\usage{
find_zero(f, upper = TRUE, start_interval, ...)
}
\arguments{
\item{f}{Function}

\item{upper}{(Boolean) Search x-values above (TRUE) or below (FALSE)
the `start_interval`}

\item{start_interval}{(numeric, 2-component vector) initial interval of x-values
in which to search for the zero. If no zero-crossing, then this function will
search additional x-values in the direction specified by `upper`}
}
\description{
Find zeros of a concave function
}
\note{
This method assumes a concave function.
}
\keyword{internal}
