% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data-wrapper.R
\name{generate_data_wrapper}
\alias{generate_data_wrapper}
\title{Generate simulated or case study data for n-of-1 Cox regression analysis.}
\usage{
generate_data_wrapper(sim = TRUE, param = NA, x_vars = NA, p = 1,
  verbose = FALSE)
}
\arguments{
\item{sim}{(Boolean) Simulate data (TRUE) or read in case study data (FALSE)}

\item{param}{(tbl) 1-row tibble of parameters, output from `create_params`}

\item{x_vars}{(Character vector) predictor variable names to use}

\item{p}{label for this parameter specification (usually the row number of the
parameter data)}

\item{verbose}{(Boolean) TRUE = print progress, FALSE = print nothing}
}
\value{
list of tibbles, one per (simulated or case study) participant.
Each row represents a participant-day.
For simulated data, each tibble has the following variables:
\describe{
  \item{id}{Study participant ID}
  \item{mig.num}{Next headache number to occur}
  \item{t0}{Observation start time, in days (always an integer)}
  \item{t1}{Observation end time, in days}
  \item{mig.ind}{Indicator for a day with new headache onset}
  \item{X001...X00`n`}{Predictor variables, `n` columns}
  \item{order}{sort order for participant-days}
  \item{p}{Parameter specification under which these data were generated}
}
For case study data, the variables are documented under `case_study_data.`
}
\description{
Generate simulated or case study data for n-of-1 Cox regression analysis.
}
