% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-final-data.R
\name{generate_final_data}
\alias{generate_final_data}
\title{Organize the predictors and event times into final data set.}
\usage{
generate_final_data(event_times, events.per.person, n.subj, x.list,
  max_time_for_person = NA, round_event_times = FALSE)
}
\arguments{
\item{event_times}{(numeric vector) actual event times}

\item{events.per.person}{(numeric vector) number of events for each subject
(length should equal n.subj)}

\item{n.subj}{(integer) number of subjects)}

\item{x.list}{(list of tbls) list of predictor data output from `generate_pred_wrapper()`}

\item{max_time_for_person}{(integer, optional) maximum number of time units.
Each subject's follow-up time will be cut off after `max_time_for_person` (or
not at all , if set to NA).}
}
\description{
Organize the predictors and event times into final data set.
}
\references{
Hendry DJ. Data generation for the Cox proportional hazards model with time-dependent covariates: A method for medical researchers. Stat Med. 2014;33:436-454.
}
\keyword{internal}
