% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-outcome.R
\name{generate_outcomes}
\alias{generate_outcomes}
\title{Generate outcomes for the proportional hazards model}
\usage{
generate_outcomes(rates, g.inv.t.min, g.inv.t.max, g.inv.t)
}
\arguments{
\item{rates}{(numeric vector) piecewise exponential event rates per time unit,
generated by `generate_pred_wrapper`}

\item{g.inv.t.min}{(numeric) minimum allowed event time, transformed by `generate_time_transform`}

\item{g.inv.t.max}{(numeric) maximum allowed event time, transformed by `generate_time_transform`}

\item{g.inv.t}{(numeric vector) time units corresponding to piecewise exponential rates, transformed by `generate_time_transform`}
}
\description{
Generate outcomes for the proportional hazards model
}
\references{
Hendry DJ. Data generation for the Cox proportional hazards model with time-dependent covariates: A method for medical researchers. Stat Med. 2014;33:436-454.
}
\keyword{internal}
