% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-predictor.R
\name{generate_pred_wrapper}
\alias{generate_pred_wrapper}
\title{Generate time-series of predictor variable(s): wrapper around `generate_predictor`}
\usage{
generate_pred_wrapper(n_events, time_units, var.types, rho, freq, beta)
}
\value{
list with columns for predictor variables and exponential rates (`rate`)
}
\description{
Generate time-series of predictor variable(s): wrapper around `generate_predictor`
}
\keyword{internal}
