% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-predictor.R
\name{generate_predictor}
\alias{generate_predictor}
\title{Generate time-series of predictor variable}
\usage{
generate_predictor(var.type, r, f, t)
}
\arguments{
\item{var.type}{(string) "b" for binary or "c" for continuous}

\item{r}{desired autocorrelation}

\item{f}{desired frequency (binary variables only)}

\item{t}{time vector (used for length only)}
}
\description{
Generate time-series of predictor variable
}
\keyword{internal}
