% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-outcome.R
\name{k}
\alias{k}
\title{Accept-reject function}
\usage{
k(x, lower_bound, upper_bound, pexp_rates, pexp_t)
}
\arguments{
\item{x}{value of simulated piecewise exponential}

\item{lower_bound}{minimum allowed event time, transformed by `generate_time_transform`}

\item{upper_bound}{maximum allowed event time, transformed by `generate_time_transform`}

\item{pexp_rates}{piecewise exponential rates}

\item{pexp_t}{time units corresponding to piecewise exponential rates, transformed by `generate_time_transform`}
}
\description{
Returns piecewise exponential density of `x` if `x` inside bounds,
otherwise returns zero.
}
\references{
Hendry DJ. Data generation for the Cox proportional hazards model with time-dependent covariates: A method for medical researchers. Stat Med. 2014;33:436-454.
}
\keyword{internal}
