% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-results-utils.R
\name{multivariable_summary}
\alias{multivariable_summary}
\title{Summary of results for multivariable setting}
\usage{
multivariable_summary(data_grouped)
}
\description{
Summary of results for multivariable setting
}
\keyword{internal}
