% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-case-study-est.R
\name{plot_cs_est}
\alias{plot_cs_est}
\title{Plot case-study log-hazard ratio estimates}
\usage{
plot_cs_est(data, vars, mle_class = FALSE, ub_display = NA)
}
\arguments{
\item{data}{Raw result from cox regression on the case study data (just read in result1.rds)}
}
\description{
Facet plot with 3 variable-rows and 4 participant-columns.
In each facet, plots the log-hazard ratio estimate and 90% confidence interval for
each of the priors and event counts.
}
\keyword{internal}
