% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-classifications.R
\name{plot_effect_class}
\alias{plot_effect_class}
\title{Plot effect classifications}
\usage{
plot_effect_class(res, univ = TRUE, vartype = "c")
}
\arguments{
\item{res}{Results from `summarise_results_wrapper(rowwise = FALSE)`}

\item{univ}{(Boolean) Univariate (TRUE) or multivariable (FALSE) results?}

\item{vartype}{(Character) For `univ = TRUE` only, predictor vartype = "c"
(Continuous) or "b" (Binary)}
}
\description{
Plot effect classifications
}
\keyword{internal}
