% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-effect-class-multivar.R
\name{plot_effect_class_multivar}
\alias{plot_effect_class_multivar}
\title{Plot effect classifications for multivariable setting}
\usage{
plot_effect_class_multivar(res)
}
\arguments{
\item{res}{Results from `summarise_results_wrapper(rowwise = FALSE)`}
}
\description{
Plot effect classifications for multivariable setting
}
\keyword{internal}
