% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-power.R
\name{plot_power}
\alias{plot_power}
\title{Plot Simulated Power}
\usage{
plot_power(data)
}
\arguments{
\item{data}{(tbl) summarised results, output from `summarise_results_wrapper`}
}
\description{
Plot Simulated Power
}
\keyword{internal}
