% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-univariate.R
\name{plot_univariate}
\alias{plot_univariate}
\title{Plot any result}
\usage{
plot_univariate(data, y = "me_90", max_y = 10, hline_intercept = 0)
}
\arguments{
\item{data}{(tbl) summarised results, output from `summarise_results_wrapper`}

\item{y}{(string) name of variable to plot}

\item{max_y}{(numeric) maximum value to plot on y-axis}

\item{hline_intercept}{(numeric) y-value at which to draw reference line}
}
\description{
Plot any result
}
\keyword{internal}
