% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-cis.R
\name{profile_ci}
\alias{profile_ci}
\title{Profile likelihood for coefficients in a \code{coxph} model}
\usage{
profile_ci(x, parm, CI = 0.95, data)
}
\arguments{
\item{x}{A \code{coxph} model.}

\item{parm}{(string) name of variable for which to evaluate CI.}

\item{CI}{nominal coverage level of \bold{C}onfidence \bold{I}nterval.}

\item{data}{(tbl) Dataset used to estimate the original model}
}
\description{
Profile likelihood for coefficients in a \code{coxph} model
}
\references{
Much of this code was taken from \code{survMisc::profLik}.
}
\keyword{internal}
