% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-results-utils.R
\name{rowwise_summary}
\alias{rowwise_summary}
\title{Rowwise Summary}
\usage{
rowwise_summary(data, threshold_null_or = 2, threshold_sig_or = 1)
}
\description{
Rowwise Summary
}
\keyword{internal}
