% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-results-utils.R
\name{se_cutoff}
\alias{se_cutoff}
\title{Constant SE cutoff to identify monotone likelihood}
\usage{
se_cutoff()
}
\description{
Constant SE cutoff to identify monotone likelihood
}
\keyword{internal}
