% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-univariate.R
\name{smry_univariate}
\alias{smry_univariate}
\title{Summarize a single variable to match `plot_univariate`}
\usage{
smry_univariate(data, var)
}
\arguments{
\item{data}{(tbl) output from `summarise_results_wrapper`}

\item{var}{(string) name of variable to summarize}
}
\description{
Summarize a single variable to match `plot_univariate`
}
\keyword{internal}
