% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-cox-data.R
\name{summarise_cox_data}
\alias{summarise_cox_data}
\title{Summarise Cox data for n-of-1 power analysis}
\usage{
summarise_cox_data(data, x_vars = NULL, sim = TRUE)
}
\description{
Summarise Cox data for n-of-1 power analysis
}
\keyword{internal}
