% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-errors.R
\name{summarise_errors}
\alias{summarise_errors}
\title{Summarise Errors}
\usage{
summarise_errors(results_dir, mv, sim)
}
\arguments{
\item{results_dir}{results_dir directory of saved RDS results from estimation step}

\item{mv}{(boolean) TRUE = multivariable simulation, FALSE = univariable}

\item{sim}{(boolean) TRUE = simulated data, FALSE = case study}
}
\description{
Summarise Errors
}
\keyword{internal}
