% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-univ-effect-class.R
\name{summarise_multi_effect_class}
\alias{summarise_multi_effect_class}
\title{Summarise effect classifications (multivariable)}
\usage{
summarise_multi_effect_class(res)
}
\arguments{
\item{res}{(tbl) Results from `summarise_results_wrapper`}
}
\description{
Summarise effect classifications (multivariable)
}
\keyword{internal}
