% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-results.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarize results}
\usage{
summarise_results(data, sim = TRUE, multivariable = FALSE,
  rowwise = FALSE, row_errors = FALSE)
}
\arguments{
\item{data}{tbl of saved results, slightly modified by `summarise_results_wrapper`}

\item{sim}{(boolean) TRUE = simulated data, FALSE = case study data}

\item{multivariable}{(boolean) TRUE = multivariable data, FALSE = univariable data}

\item{rowwise}{(boolean) TRUE = just return the rowwise summarized data}

\item{row_errors}{(boolean) TRUE = just return the error classifications for each row}
}
\description{
Summarize results
}
\keyword{internal}
