% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-results-wrapper.R
\name{summarise_results_wrapper}
\alias{summarise_results_wrapper}
\title{Summarize results of model estimation.}
\usage{
summarise_results_wrapper(results_dir, mv, sim, rowwise = FALSE,
  row_errors = FALSE)
}
\arguments{
\item{results_dir}{directory of saved RDS results from estimation step}

\item{mv}{(boolean) TRUE = multivariable simulation, FALSE = univariable}

\item{sim}{(boolean) TRUE = simulated data, FALSE = case study}

\item{rowwise}{(boolean) TRUE = just return the rowwise summarized data}

\item{row_errors}{(boolean) TRUE = just return the error classifications for each row}
}
\description{
Summarize results of model estimation.
}
