% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-univ-effect-class.R
\name{summarise_univ_effect_class}
\alias{summarise_univ_effect_class}
\title{Summarise effect classifications (univariable)}
\usage{
summarise_univ_effect_class(res, vartype)
}
\arguments{
\item{res}{(tbl) Results from `summarise_results_wrapper`}

\item{vartype}{(character) Variable type to use ("b" for binary or "c" for continuous)}
}
\description{
Summarise effect classifications (univariable)
}
\keyword{internal}
